(fix): update slot id parsing

+++ pkg/signing/pkcs11/pkcs11uri_test.go
@@ -158,7 +158,7 @@ func TestURI_SlotID(t *testing.T) {
 	tests := []struct {
 		name      string
 		uri       string
-		wantSlot  int
+		wantSlot  int64
 		wantErr   bool
 		errString string
 	}{
